+++ packages/core/src/tracing/getActiveSpan.ts
@@ -1,0 +1,2 @@
+import type { Span } from '@sentry/types';
+import { getCurrentScope } from '../currentScopes';
