+++ packages/core/src/tracing/index.ts
@@ -4,7 +4,7 @@ export type { BeforeFinishCallback } from './idletransaction';
 export { SentrySpan } from './sentrySpan';
 export { Transaction } from './transaction';
 // eslint-disable-next-line deprecation/deprecation
-export { getActiveTransaction } from './utils';
+export { getActiveTransaction, getActiveSpan } from './utils';
 // eslint-disable-next-line deprecation/deprecation
 export { SpanStatus } from './spanstatus';
 export {
@@ -13,7 +13,6 @@ export {
 } from './spanstatus';
 export type { SpanStatusType } from './spanstatus';
 export {
-  getActiveSpan,
   startSpan,
   startInactiveSpan,
   startSpanManual,
